backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.2
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.3
dnspython==2.6.1
